+++ sources/modules/Calendar/CalendarPostModule.class.php
@@ -45,7 +45,7 @@ class Calendar_Post_Module extends ElkArte\sources\modules\Abstract_Module
 		}
 		$context['make_event'] = self::$_make_event;
 		$context['cal_minyear'] = $modSettings['cal_minyear'];
-		$context['cal_maxyear'] = date('Y') + $modSettings['cal_limityear'];
+		$context['cal_maxyear'] = (int) date('Y') + (int) $modSettings['cal_limityear'];
 
 		if (self::$_make_event)
 			return array(
@@ -232,7 +232,7 @@ class Calendar_Post_Module extends ElkArte\sources\modules\Abstract_Module
 			if ($context['event']['month'] < 1 || $context['event']['month'] > 12)
 				throw new Elk_Exception('invalid_month', false);
 
-			if ($context['event']['year'] < $modSettings['cal_minyear'] || $context['event']['year'] > date('Y') + $modSettings['cal_limityear'])
+			if ($context['event']['year'] < $modSettings['cal_minyear'] || $context['event']['year'] > (int) date('Y') + (int) $modSettings['cal_limityear'])
 				throw new Elk_Exception('invalid_year', false);
 
 			// Get a list of boards they can post in.
